backend/requirements.txt
Django==3.0.6
asgiref==3.2
channels
channels-redis
djangorestframework
django-cors-headers
django-grappelli==2.13.3
drf-yasg
celery
django-celery-results
django-celery-beat
django-filter==2.2.0
django-extensions 
django-image-cropping
easy-thumbnails
python-telegram-bot==12.0.0b1
channels
djangorestframework-oauth
django-oauth-toolkit